pyproject.toml
[project]
name = "wallet-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.5",
    "dotenv>=0.9.9",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "paystack>=1.5.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]
